refactor(mux): rehome Zellij cache cleanup
Keep serialized-session layout knowledge inside the Zellij adapter.
